Remove unused CSS files and update styles for improved layout and responsiveness
