+++ package.json
@@ -20,12 +20,12 @@
     "framer-motion": "^12.0.6",
     "lucide-react": "^0.471.0",
     "next-themes": "^0.4.4",
-    "pdfjs-dist": "^3.11.174",
+    "pdfjs-dist": "^4.10.38",
     "prettier": "^3.4.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^5.4.0",
-    "react-pdf": "^7.5.1",
+    "react-pdf": "^9.2.1",
     "react-router-dom": "^7.1.1",
     "react-slick": "^0.30.3",
     "slick-carousel": "^1.8.1",
