--- src/components/Aboutus/AboutUsInside/Mission/Mission.css
@@ -1,11 +1,0 @@
-.mission-text {
-    font-weight: bold;
-    margin: 0.5rem;
-}
-.mission{
-    margin-top: 2rem;
-    margin-left: 3rem;
-}
-.mission-heading{
-    margin: 0.5rem;
-}
