+++ src/components/Aboutus/AboutUsInside/Motto/Motto.tsx
@@ -1,135 +1,23 @@
-import { useEffect, useRef, useState, RefObject } from "react";
-import { motion } from "framer-motion";
-import "./Motto.css"
+import { Box, Em, Heading } from '@chakra-ui/react';
+import React from 'react';
+import TrueFocus from '@/components/ui/TextFocus/textFocus';
 
-interface TrueFocusProps {
-    sentence?: string;
-    manualMode?: boolean;
-    blurAmount?: number;
-    borderColor?: string;
-    glowColor?: string;
-    animationDuration?: number;
-    pauseBetweenAnimations?: number;
-}
-
-interface FocusRect {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-}
-
-const TrueFocus: React.FC<TrueFocusProps> = ({
-    sentence = "COME CLICK CREATE",
-    manualMode = false,
-    blurAmount = 5,
-    borderColor = "#00D8FF",
-    glowColor = "#00D8FF",
-    animationDuration = 0.5,
-    pauseBetweenAnimations = 1,
-}) => {
-    const words = sentence.split(" ");
-    const [currentIndex, setCurrentIndex] = useState<number>(0);
-    const [lastActiveIndex, setLastActiveIndex] = useState<number | null>(null);
-    const containerRef: RefObject<HTMLDivElement> = useRef(null);
-    const wordRefs: React.MutableRefObject<(HTMLSpanElement | null)[]> = useRef([]);
-    const [focusRect, setFocusRect] = useState<FocusRect>({ x: 0, y: 0, width: 0, height: 0 });
-
-    useEffect(() => {
-        if (!manualMode) {
-            const interval = setInterval(() => {
-                setCurrentIndex((prev) => (prev + 1) % words.length);
-            }, (animationDuration + pauseBetweenAnimations) * 1000);
-
-            return () => clearInterval(interval);
-        }
-    }, [manualMode, animationDuration, pauseBetweenAnimations, words.length]);
-
-    useEffect(() => {
-        if (currentIndex === null || currentIndex === -1) return;
-
-        if (!wordRefs.current[currentIndex] || !containerRef.current) return;
-
-        const parentRect = containerRef.current.getBoundingClientRect();
-        const activeRect = wordRefs.current[currentIndex]!.getBoundingClientRect();
-
-        setFocusRect({
-            x: activeRect.left - parentRect.left,
-            y: activeRect.top - parentRect.top,
-            width: activeRect.width,
-            height: activeRect.height,
-        });
-    }, [currentIndex, words.length]);
-
-    // Handlers for manual mode (hover)
-    const handleMouseEnter = (index: number) => {
-        if (manualMode) {
-            setLastActiveIndex(index);
-            setCurrentIndex(index);
-        }
-    };
-
-    const handleMouseLeave = () => {
-        if (manualMode) {
-            setCurrentIndex(lastActiveIndex ?? 0);
-        }
-    };
-
+const Motto: React.FC = () => {
     return (
-        <div className="focus-container" ref={containerRef}>
-            {words.map((word, index) => {
-                const isActive = index === currentIndex;
-                return (
-                    <span
-                        key={index}
-                        ref={(el) => (wordRefs.current[index] = el)}
-                        className={`focus-word ${manualMode ? "manual" : ""} ${isActive && !manualMode ? "active" : ""
-                            }`}
-                        style={{
-                            filter:
-                                manualMode
-                                    ? isActive
-                                        ? `blur(0px)`
-                                        : `blur(${blurAmount}px)`
-                                    : isActive
-                                        ? `blur(0px)`
-                                        : `blur(${blurAmount}px)`,
-                            transition: `filter ${animationDuration}s ease`,
-                            "--border-color": borderColor,
-                            "--glow-color": glowColor,
-                        } as React.CSSProperties}
-                        onMouseEnter={() => handleMouseEnter(index)}
-                        onMouseLeave={handleMouseLeave}
-                    >
-                        {word}
-                    </span>
-                );
-            })}
-
-            <motion.div
-                className="focus-frame"
-                animate={{
-                    x: focusRect.x,
-                    y: focusRect.y,
-                    width: focusRect.width,
-                    height: focusRect.height,
-                    opacity: currentIndex >= 0 ? 1 : 0,
-                }}
-                transition={{
-                    duration: animationDuration,
-                }}
-                style={{
-                    "--border-color": borderColor,
-                    "--glow-color": glowColor,
-                } as React.CSSProperties}
-            >
-                <span className="top-left corner"></span>
-                <span className="top-right corner"></span>
-                <span className="bottom-left corner"></span>
-                <span className="bottom-right corner"></span>
-            </motion.div>
-        </div>
+        <Box>
+            <Heading as="h2" size="3xl" color="white" mt={8} mb={4} >
+                Our <Em>Motto</Em>
+            </Heading>
+            <TrueFocus
+                sentence="COME CLICK CREATE"
+                manualMode={false}
+                blurAmount={5}
+                borderColor="red"
+                animationDuration={2}
+                pauseBetweenAnimations={1}
+            />
+        </Box>
     );
 };
 
-export default TrueFocus;
+export default Motto;
