+++ src/components/Aboutus/AboutUsInside/TeamImage/TeamImage.tsx
@@ -3,12 +3,11 @@ import { Box, Image } from '@chakra-ui/react';
 
 const TeamImage = () => {
   return (
-    <Box mb={12}>
+    <Box>
       <Image
         src="https://images.pexels.com/photos/30336520/pexels-photo-30336520/free-photo-of-group-gathering-in-traditional-moroccan-interior.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1"
         alt="PixElation Team"
         borderRadius="lg"
-        width="90%"
       />
     </Box>
   );
