--- src/components/OurWork/WorkInside/Header/workheader.css
@@ -1,7 +1,0 @@
-.header-text {
-    text-shadow: 4px 4px 11px rgba(180, 180, 180, 0.6);
-    padding-bottom: 4rem; 
-    font-family: 'Gloock', serif;
-  font-weight: 600;
-  font-style: normal;
-}
