+++ src/components/OurWork/WorkInside/Header/workheader.tsx
@@ -2,8 +2,8 @@ import { Heading, Box , Em} from '@chakra-ui/react';
 
 const Header = () => {
   return (
-    <Box mx={12} my={6}>
-      <Heading as="h1" fontSize="5xl" mb={12} color="white" className='header-text'>
+    <Box>
+      <Heading as="h1" fontSize="4xl" color="white" className='header-text'>
         Our <Em>Work</Em>
       </Heading>
     </Box>
