+++ src/components/ui/TextFocus/textFocus.css
@@ -1,38 +1,37 @@
-  .focus-container {
+.focus-container {
     position: relative;
     display: flex;
     gap: 1em;
-    justify-content: center;
+    justify-content: start;
     align-items: center;
     flex-wrap: wrap;
-  }
-  
-  .focus-word {
-    margin-top: 2rem;
-    margin-bottom: 2rem;
+}
+
+/* Words */
+.focus-word {
     position: relative;
-    font-size: 3rem;
+    font-size: 2rem;
     font-weight: 900;
     cursor: pointer;
     transition:
-      filter 0.3s ease,
-      color 0.3s ease;
-  }
-  
-  .focus-word.active {
+        filter 0.3s ease,
+        color 0.3s ease;
+}
+
+.focus-word.active {
     filter: blur(0);
-  }
-  
-  .focus-frame {
+}
+
+.focus-frame {
     position: absolute;
     top: 0;
     left: 0;
     pointer-events: none;
     box-sizing: content-box;
     border: none;
-  }
-  
-  .corner {
+}
+
+.corner {
     position: absolute;
     width: 1rem;
     height: 1rem;
@@ -40,33 +39,32 @@
     filter: drop-shadow(0px 0px 4px var(--border-color, #fff));
     border-radius: 3px;
     transition: none;
-  }
-  
-  .top-left {
+}
+
+.top-left {
     top: -10px;
     left: -10px;
     border-right: none;
     border-bottom: none;
-  }
-  
-  .top-right {
+}
+
+.top-right {
     top: -10px;
     right: -10px;
     border-left: none;
     border-bottom: none;
-  }
-  
-  .bottom-left {
+}
+
+.bottom-left {
     bottom: -10px;
     left: -10px;
     border-right: none;
     border-top: none;
-  }
-  
-  .bottom-right {
+}
+
+.bottom-right {
     bottom: -10px;
     right: -10px;
     border-left: none;
     border-top: none;
-  }
-  
+}
