--- src/pages/About/AboutInside/AboutUs.css
@@ -1,5 +1,0 @@
-.aboutus {
-    max-height: 100vh;
-    margin-left: 12px;
-    margin-right: 6px;
-}
